Add statements that show how JavaScript can be confusing
hatPrice has been expressed as a number, while the bootsPrice value is a
string of characters, enclosed in double quotes. In most languages,
performing operations on different types would be an error. JavaScript
is different; comparing a string and a number succeeds, but trying to
total the values actually concatenates them.