Demonstrate type coercion when using the logical OR operator
In addition to null and undefined, the logical OR operator will also
coerce the number value 0 (zero), the empty string value (""), and the
special NaN number value to false. These values, in addition to the
false value, are collectively known as the JavaScript "falsy" values and
cause a lot of confusion.

This commit shows he logical OR operator uses the fallback value when
the taxRate variable is assigned zero and produces the following output:

Hat price: 100
Boots price: 100
Tax rate: 10%
Tax rate: 10%

The code doesn't differentiate between an unassigned value and the zero
value, which can be a problem when zero is a required value. In this
code, it is impossible to set a tax rate of zero, even though this is a
legitimate rate.