Invoke sumPrices in different ways and write the result and its type
The first function call includes a string argument, which causes all of
the function's parameters to be converted to string values and
concatenated, meaning that the function returns the string value
100100undefined.

The second function call uses three number values, which are added
together and produce the number result 600. The final function call uses
number arguments but doesn't provide a third value, which causes an
undefined parameter. JavaScript coalesces undefined to the special
number value NaN (meaning not a number). The result of addition that
includes NaN is NaN, which means that the type of the result is number
but the value isn't useful and is unlikely to be what was intended.