Invoke a stand-alone function using the this keyword
When the code is saved and executed, the following output is produced:

```
Hello, It is sunny today
```

JavaScript defines a global object, which can be assigned values that
are available throughout an application. The global object is used to
provide access to the essential features in the execution environment,
such as the document object in browsers that allows interaction with the
Document Object Model API.

Values assigned names without using the let, const, or var keyword are
assigned to the global object. The statement that assigns the string
value Hello creates a variable in the global scope. When the function is
executed, this is assigned the global object, so reading this.greeting
returns the string value Hello, explaining the output produced by the
application.