Invoke the function without using the convencience syntax
The standard way to invoke a function is to use parentheses that contain
arguments, but in JavaScript, this is a convenience syntax that is
translated into the call method.

Functions are objects, which means they define methods, including the
call method. It is this method that is used to invoke a function behind
the scenes. The first argument to the call method is the value for this,
which is set to the global object. This is the reason that this can be
used in any function and why it returns the global object by default.

The name of the global object changes based on the execution
environment. In code executed by Node.js, global is used, but window or
self may be required in browsers.

At the time of writing, there is a proposal to standardize the name
global, but it has yet to be adopted universally.