Invoke the function when it is assigned to a function
Care is required because this is set differently if the function is
accessed outside of its object, which can happen if the function is
assigned to a variable.

Functions can be used like any other value, including assigning them to
variables outside of the object in which they were defined. If the
function is invoked through the variable, then this will be set to the
global object. This often causes problems when functions are used as
arguments to other methods or as callbacks to handle events, and the
effect is that the same function will behave differently based on how it
is invoked.