Use the function's bind method, to set the value for this
The bind method returns a new function that will have a persistent value
for this when it is invoked. The function returned by the bind method is
used to replace the original method, ensuring consistency when the
writeMessage method is invoked.

Using bind is awkward because the reference to the object isn't
available until after it has been created, which leads to a two-step
process of creating the object and then calling bind to replace each of
the methods for which a consistent this value is required.