Use an arrow function to understand this in different contexts
Arrow functions don't have their own this value and inherit the closest
value of this they can find when they are executed.

The getWriter function is a regular function that returns an arrow
function as its result. When the arrow function returned by getWriter is
invoked, it works its way up its scope until it locates a value for
this. As a consequence, the way that the getWriter function is invoked
determines the value of this for the arrow function.

When the getWriter method is invoked through myObject it means that the
value of this will be set to myObject. When the arrow function is
invoked, it finds a value of this from the getWriter function. The
result is that when the getWriter method is invoked through myObject,
the value of this in the arrow function will be myObject, and the
this.greeting expression in the template string will be Hi, there.

The statements in the second set treat getWriter as a stand-alone
function, so this will be set to the global object. When the arrow
function is invoked, the this.greeting expression will be Hello.