Specify the lib compiler option and add types so Map is supported
The set of types selected includes the standard types for the version of
JavaScript selected by the target property, the dom setting (which
provides access to the console object), and the ES2015 collections
feature.

It is important to think through the implications of using the lib
configuration setting because it just tells the TypeScript compiler that
the runtime for the application can be relied on to support a specific
set of features, such as the Map in this case. The compiler can adapt
the JavaScript it generates for different language features, but that
doesn't extend to objects like collections. Changing the lib setting
tells the compiler that there will be a nonstandard set of features
available when the compiled JavaScript is executed, and it is your
responsibility to ensure this is the case, either because you know more
about the runtime than the compiler or because the application uses a
polyfill such as core-js (https://github.com/zloirock/core-js).

The Node.js version 18.14 supports most of the recent JavaScript
features and can be relied on to have Map.