Add a calc module and use it in index
The TypeScript compiler uses the target configuration property to select
the approach taken to deal with modules. When the target is es5, it uses
the CommonJS module style, which was the result of an earlier attempt to
introduce a module standard before ECMAScript modules were widely
adopted. The Node.js runtime supports the commonjs module system, which
is why the code generated by the TypeScript compiler executes without
problems.