Specify a module format and include file extension in import statement
This requirement is an oddity because it requires the file extension of
the JavaScript file that is produced by the compiler to be specified in
the TypeScript file. The TypeScript development team has adopted the
principle of not rewriting paths in import statements, which means that
the path component of the import statement, which specifies the file
name, must be written for the JavaScript runtime and not the TypeScript
compiler.