Add a myVar variable and set its value to different types
The biggest obstacle that JavaScript presents to programmers who are
used to other languages is that values have types instead of variables.

The type of the variable named myVar changes based on the value assigned
to it. The JavaScript typeof keyword can be used to determine a type.

The first statement defines the variable without assigning a value,
which means that its type is undefined. A variable whose type is
undefined will always have a value of undefined, which can be seen in
the output.

The value 12 is a number, and as soon as the value is assigned, the data
type of the variable changes. The value Hello is a string, and the value
false is a boolean; you can see the data type as each value is assigned
to the variable. You don't need to tell JavaScript the data type, which
it automatically infers from the value.