Use toFixed to format the number as a string
This is legal JavaScript and will be dealt with by type coercion. In
this case, the string value will be converted to a number, and the
outcome will be either the division of two number values or NaN if the
string value cannot be converted.

In TypeScript, automatic type coercion is restricted, and the compiler
reports an error instead of trying to convert values.