Disable implicit any types
Explicitly using any provides an escape hatch from type checking, which
can be useful when applied cautiously. Allowing the compiler to use any
implicitly creates gaps in type checking that you may not even notice
and that can undermine the benefit of using TypeScript.

It is good practice to disable the implicit use of any by setting the
compiler's noImplicityAny setting. (The implicit use of any is also
disabled when you enable the strict compiler setting.)