Declare union types explicitly on function return values
It is important to understand that a type union is handled as a type in
its own right, whose features are the intersection of the individual
types. This means that the type of the taxNumber, for example, is
string | number and not number, even though the calculateTax function
returns a number when the boolean argument is false.

You can only use the properties and methods defined by all the types in
the union, which can be useful for complex types but is limited by the
small common API presented by primitive values. The only method shared
by the number and string types that are used in the union is the
toString method.

This means that the other methods defined by the number and string types
cannot be used, and the use of the toFixed and charAt methods produces
the compiler errors.