Assert the return value to an unexpected type
The compiler checks that the type used in an assertion is expected. When
using an assertion from a type union, for example, the assertion must be
to one of the types in the union.

The new type assertion in this commit tells the compiler to treat a
string | number value as a boolean. The compiler knows that boolean is
not one of the types in the union and produces the following error when
the code is compiled:

src/index.ts(29,18): error TS2352: Conversion of type 'string | number'
 to type 'boolean' may be a mistake because neither type sufficiently
 overlaps with the other. If this was intentional, convert the
 expression to 'unknown' first.
  Type 'number' is not comparable to type 'boolean'.