Force the assertion by first asserting to any
In most situations, you should review the data types and the type
assertion and correct the problem by expanding the type union or
asserting to a different type. However, you can force the assertion and
override the compiler's warning by first asserting to any and then to
the type you require.

This additional step prevents the compiler from warning about the change
and treats the result from the function as a boolean value. However
assertions only affect the type-checking process and do not perform type
coercion.

The result produced by the function has been described to the compiler
as the string | number union and asserted as a boolean. But when the
code is executed, the function produces a number, whose value is written
to the console.