Use a type guard to test for a specific type
To test a type, the typeof keyword is applied to a value, producing a
string that can be compared to the names of the primitive JavaScript
types, such as number and boolean.

The typeof keyword can be used only with the JavaScript primitive types.
A different approach is required to differentiate between objects.

The compiler doesn't implement the typeof keyword, which is part of the
JavaScript specification. Instead, the compiler trusts that the
statements in the conditional block will be executed at runtime only if
the value being tested is of the specified type. This knowledge allows
the compiler to treat the value as the type being tested.