Use a switch statement instead of an if...else block
The compiler is adept at recognizing type guard statements, even when
they are not in a conventional if...else block. This commit produces the
same result but uses a switch statement to differentiate between types.
Within each block, the compiler treats taxValue as though it has been
defined with only the type selected by the case statement.