Use a never type to indicate code that should never execute
TypeScript provides the never type for situations where a type guard has
dealt with all of the possible types for a value. Here, for example, the
switch statement is a type guard for the number and string types, which
are the only types that will be returned in the string | number union
from the function. Once all the possible types have been handled, the
compiler will only allow a value to be assigned to the never type.

Something has gone wrong if execution reaches the default clause of the
switch statement, and TypeScript provides the never type to ensure you
can't accidentally use a value once type guards have been used to
exhaustively narrow a value to all of its possible types.