USe the unknown type
TypeScript supports the unknown type, which is a safer alternative to
any. An unknown value can be assigned only any or itself unless a type
assertion or type guard is used.

An unknown value can't be assigned to another type without a type
assertion, so the compiler produces the following error when it compiles
the code:

src/index.ts(54,5): error TS2322: Type 'unknown' is not assignable to
 type 'number'.