Use a nullable type in the calculateTax function if the amount is zero
The change to the calculateTax shows a typical use of null, where it is
used as a result if the value of the amount parameter is zero,
indicating an invalid condition. The result type for the function and
the type of the taxValue variable are string | number. But, in
JavaScript, changing the value assigned to a variable can change its
type, and that is what happens in the example: the call to the
calculateTax function returns null, which changes the taxValue type to
null. When the type guard statements inspect the type of the variable,
they fail to narrow its type to one of those in the string | number
union and instead reaches the default clause of the switch statement and
is assigned to the never type.

Under normal circumstances, the compiler will report an error if a value
of one type is assigned to a variable of a different type, but the
compiler remains silent because it allows null and undefined to be
treated as values for all types.