Use a non-null assertion to tell the compiler that taxValue is not null
A non-null value is asserted by applying the ! character after the
value. The assertion tells the compiler that the result from the
calculateTax function will not be null, which allows it to be assigned
to the taxValue variable, whose type is string | number.

Unions present the intersection of the API of each type. The null and
undefined values don't present any properties or methods, which means
that values for nullable type unions can't be used directly, even if the
non-null types have an intersection of useful properties or methods. A
non-null assertion tells the compiler that a value isn't null, which
removes null from the type union and allows the intersection of the
other types to be used.