Use a non-null type guard instead of a non-null assertion
Even though the nonNullTaxValue variable is never used, the compiler
knows that the test for null values means that the value can be treated
as the non-nullable string | number union type with the if code block.

The compiler also knows that taxValue can be null only in the else code
block.