Redefine the calculateTax function
Many languages support function overloading, which allows multiple
functions to be defined with the same name as long as they have
different numbers of parameters or if the parameters have different
types. If you are used to this style of programming, this code looks
perfectly normal, and you will assume the second calculateTax function
builds on the first calculateTax function to apply a discount.

JavaScript doesn't support function overloading, and when you define two
functions with the same name, the second function replaces the first,
regardless of the function's parameters. The number of arguments used to
call a function is not important in JavaScript — if there are more
parameters than arguments, then the extra parameters are undefined. If
there are more arguments than parameters, the function can either ignore
them or use the special arguments value, which provides access to all
the arguments used to invoke the function. If this code were executed,
the first calculateTax function would be ignored, and the second
function would be invoked, but without a value for the second parameter.
When the function is executed, it would invoke itself repeatedly, until
the call stack becomes exhausted and an error is produced.