Consolidate the calculateTax functionality into a single function
The practical effect of not being able to overload functions is that
different names must be used (such as calculateTax and
calculateTaxWithDiscount, for example) or a single function adapts its
behavior based on its parameters.

Here the approach of consolidating the functionality into a single
function has been taken, which usually only works well for simpler
tasks. For more complex groups of features using different names is
usually a better approach.