Enable the noImplicitReturns compiler option
The type guard used to filter out null values means that there is a path
through the function's code that doesn't reach a return statement and so
the function will return a number if the amount parameter isn't null and
will return undefined if the amount parameter is null. With the
strictNullChecks compiler option enabled, the compiler has inferred the
result type to be number | undefined. To prevent implicit returns,
enable the noImplicitReturns compiler setting.