Add a type union and use a type guard on the result
The function will always return a number result if the amount parameter
is a number parameter and will always return null if amount is null. The
user of the function has to use a type guard on the result to remove
null values, even though when the value is a number it will always
produce a number result.