Overload the function types so a type guard doesn't need to be used
The type overloads replace the function definition as the type
information used by the TypeScript compiler, which means that only those
combinations of types can be used. When the function is invoked, the
compiler can determine the result type based on the type of the
arguments provided, allowing the taxValue variable to be defined as a
number and removing the need for the type guard to pass on the result to
the writeValue function. The compiler knows that taxValue can only be a
number and doesn't require the type to be narrowed.

NOTE This is not the function overloading supported by languages such as
C# and Java. Only the type information is overloaded by this feature for
the purposes of type checking. There is only one implementation of the
function, which is still responsible for dealing with all the types used
in the overloads.