Create an empty array and then add the values to the array
There are no initial values for the compiler to use when selecting the
type for the prices array. The only option available to the compiler is
to use any since it has no other information to work with.

The effect of allowing the compiler to infer the type of the empty array
is to create a gap in the type-checking process. The code works because
the JavaScript multiplication operator coerces string values to number
values automatically. This can be useful behavior, but it is likely to
be used accidentally, and it is for this reason that you should use
explicit types.