Use tuples to describe the names and prices
Tuples are defined using square brackets containing the types for each
element. Tuples must be defined with type annotations; otherwise, the
compiler will assume that a regular array with a type that is the union
of each value used during initialization. Without the type annotation
the compiler would assume that the type of the value assigned to the hat
variable is [string | number], which would denote a variable-length
array in which every element can be either a string or number value.