Use enum and number values
The compiler enforces type checking for enums, which means that you will
receive an error if you try to compare values from different enums, even
when they have the same underlying number value. Enums provide an array-
indexer style syntax that can be used to get the name of a value.