Define a const enum
The TypeScript compiler creates an object that provides the
implementation for an enum. In some applications, the performance impact
of using the object can be a problem, and a different approach can be
used instead.

To prevent the compiler from using an object to implement an enum, the
const keyword can be used when the enum is defined in the TypeScript
file. Const enums are more restrictive than regular enums, and all of
the values must be assigned constant expressions. The simplest way to do
this is to allow the compiler to assign values or to explicitly assign
values yourself.