Add a DogOwner interface and change Customer to implement multiple interfaces
A class can implement multiple interfaces only if there are no
overlapping properties with conflicting types. For example, if the
Person interface defined a string property named id and if the DogOwner
interface defined a number property with the same name, the Customer
class would not be able to implement both interfaces because there is no
value that could be assigned to its id property that could represent
both types.