Add a Product interface and a type guard using the getDetails property
There is no JavaScript equivalent to interfaces, and no details of
interfaces are included in the JavaScript code generated by the
TypeScript compiler. This means that the instanceof keyword cannot be
used to narrow interface types, and type guarding can be done only by
checking for one or more properties that are defined by the interface.