Remove the restriction on generic type for DataCollection and add a filter method
Notice that the filter method's generic type parameter, named V, is
defined with the extend keyword, telling the compiler that it can only
accept types that can be assigned to the class generic type T, which
prevents the compiler from treating V as any.

This example doesn't compile and produces the following error message:

src/index.ts(35,58): error TS2693: 'V' only refers to a type, but is
being used as a value here.

No JavaScript feature is equivalent to generic types, so they are
removed from the TypeScript code during the compilation process, which
means that there is no information available at runtime to use generic
types with the instanceof keyword.