Use the keyof keyword AKA the index type query operator
The keyof keyword, known as the index type query operator, returns a
union of the property names of a type, using the literal value type
feature.

The type annotation for the myVar variable is keyof Product, which will
be the union of the property names defined by the Product class. The
result is that myVar can be assigned only the string values name and
price because these are the names of the only two properties defined by
the Product class in the dataTypes.ts file. Assigning any other value to
myVar produces a compiler error.