Constrain a generic type parameter using the keyof keyword
The getValue function has a type parameter K that is constrained using
typeof T, which means that K can be the name of only one of the
properties defined by T, regardless of the type used for T when the
function is invoked. When the getValue function is used with a Product
object, the keyname parameter can be only name or price. And when the
getValue function is used with an Employee object, the keyname parameter
can be only name or role. In both cases, the keyname parameter can be
used to safely get or set the value of the corresponding property from
the Product or Employee object.