Explicitly specify the generic type arguments
It can appear as though the property is specified twice, but name has
two different uses in the modified statement. As a generic type
argument, name is a literal value type that specifies one of the keyof
Product types and is used by the TypeScript compiler for type checking.
As a function argument, name is a string value that is used by the
JavaScript runtime when the code is executed.