Use the indexed access operator to get the type for various properties
The indexed access operator is expressed using square brackets following
a type so that Product["price"], for example, is number, since that is
the type of the price property defined by the Product class. The indexed
access operator works on literal value types.

The keyof Product expression returns a literal value type union with the
property names defined by the Product class, "name" | "price". The
indexed access operator returns the union of the types of those
properties, such that Product[keyof Product] is string | number, which
is the union of the types of the name and price properties.

The types returned by the indexed access operator are known as lookup
types.

The indexed access operator is most commonly used with generic types,
which allows property types to be handled safely even though the
specific types that will be used are unknown.