Use the indexed access operator in the getValue function
The indexed access operator, T[K], tells the compiler that the result of
the getValue function will have the type of the property whose name is
specified by the keyof type argument, leaving the compiler to determine
the result types based on the generic type arguments used to invoke the
function. For the Product object, that means a name argument will
produce a string result, and a price argument will produce a number
result.