Use an index type for the Collection<T> class
Using an index type allows the Collection<T> class to store any type of
object and not just those that define a name property. The class uses an
index type query to restrict the propertyName constructor property to
the names of the properties defined by the generic type parameter T,
providing the key by which objects can be stored in the Map.