Update the type mapping to change both the name and type of the properties
The AllowStrings type is created with a mapping that creates a type
union between string and the property's original type, like this:

...
[P in keyof Product] : Product[P] | string
...

The result is a type that is equivalent to this type:

type AllowStrings = {
    name: string;
    price: number | string;
}

The ChangeNames type is created with a mapping that alters the name of
each property by adding Property.

...
[P in keyof Product as `${P}Property`] : Product[P]
...

The as keyword is combined with an expression that defines the property
name. In this case, a template string is used to modify the existing
name, with the result that is equivalent to this type:

type ChangeNames = {
    nameProperty: string;
    priceProperty: number;
}