Use mapped types to change property optionality and mutability
A question mark (the ? character) is placed after the name selector to
make the properties in the mapped type optional, and a minus sign and a
question mark (the -? characters) are used to make properties required.
Properties are made read-only and read-write by preceding the name
selector with readonly and -readonly.

Mapped types change all the properties defined by the type they
transform so that the type produced by MakeOptional<T> when applied to
the Product class, for example, is equivalent to this type:

type optionalType = {
    name?: string;
    price?: number;
}

The types produced by mappings can be fed into other mappings, creating
a chain of transformations. The type produced by the MakeOptional<T>
mapping is then transformed by the MakeRequired<T> mapping, the output
of which is then fed to the MakeReadOnly<T> mapping and then the
MakeReadWrite<T> mapping. The result is that properties are made
optional and then required and then read-only and, finally, read-write.