Use a a generic type parameter in the mapped type to select specific properties by name
The SelectProperties mapping defines an additional generic type
parameter named K that is restricted using keyof so that only types that
correspond to properties defined by the type parameter T can be
specified. The new type parameter is used in the mapping's name
selector, with the result that individual properties can be selected for
inclusion in the mapped type, like this:

...
let p1: SelectProperties<Product, "name"> = { name: "Kayak" };
...

This mapping selects the name property defined by the Product class.
Multiple properties can be expressed as a type union, and TypeScript
provides the built-in Pick<T, K> mapping that performs the same role.

...
let p2: Pick<Product, "name"> = { name: "Kayak" };
...

The Pick mapping specifies the keys that are to be kept in the mapped
type. The Omit mapping works in the opposite way and excludes one or
more keys.

...
let p3: Omit<Product, "price"> = { name: "Kayak"};
...

The result of all three mappings is the same.