Experiment with conditional types
Conditional types have a generic type parameter and a ternary expression
that selects a result type. A conditional type is a placeholder for one
of its result types, which isn't chosen until the generic type parameter
is used, which allows the expression to be evaluated using one of the
result types selected.