Define a function that accepts an array or a single object
This code won't compile because the generic parameters don't correctly
capture the relationship between the types. If the getValue function
receives an array through the data parameter, it returns the value of
the property specified by the propName parameter for the first item in
the array. If the function receives a single object through data, then
it returns the propName value for that object. The propName parameter is
constrained using keyof, which is a problem when an array is used
because keyof returns a union of the property names defined by the
JavaScript array object and not the properties of the type contained in
the array