Reconfigure the way the TypeScript compiler resolves dependencies on modules
The paths property is used to specify locations that the TypeScript
compiler will use as it tries to resolve import statements for modules.
The configuration tells the compiler to look for all packages in a
folder called types. The cts, mts and ts file extensions have been
specified, which is important because type declarations have to
match the module formats of the package they are applied to.

When the paths property is used, the baseUrl property must also be
specified, and the value used tells the compiler that the location
specified by the path property can be found in the same folder as the
tsconfig.json file.