Add an abstractDataSource.ts file to the src/data directory
Notice that the import statements do not include the js file extensions.
The requirement for file extensions is specific to Node.js and their
interpretation of the JavaScript language specification. When a browser
is used to execute the compiled JavaScript code the file extensions need
to be omitted.

The AbstractDataSource class uses the JavaScript Promise features to
fetch data in the background and uses the async/await keywords to
express the code that depends on those operations. The class invokes the
abstract loadProducts method in the constructor, and the getProducts and
getCategories methods wait for the background operation to produce data
before returning any responses.