Update index.ts to call the LocalDataSource and output to the console
When the changes to the index.ts file are saved, the code will be
compiled, and the chain of import statements is resolved to include all
the JavaScript required by the application in the webpack bundle. A
browser reload will be triggered, and the output will be displayed in
the browser's JavaScript console.